config: add AM_PROG_CC_C_O for per-target compilation flags
Per-target compilation flags (theatre200_drv_la_CFLAGS) are required
when multiple targets which require different compiler flags,
are build in the same makefile.

Automake issues a command with -c and -o flags which not all compilers
support. The object fles are prefixed with theatre200_drv_la.
The macro AM_PROG_CC_C_O must then be used to provide this feature
on compilers that do not have it. If not, a warning is issued at make time.

This macros checks for compiler support and if missing, uses a "compile"
script it generates in the package root directory.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>