dri2: Reference count DRI2 buffers
When a client calls ScheduleSwap we set up a kernel callback when the
relevent vblank event occurs.  However, it's possible for the client
to go away between calling ScheduleSwap and the vblank event,
resulting in the buffers being destroyed before they're passed to
radeon_dri2_frame_event_handler.

Add reference-counting to the buffers and take a reference in
radeon_dri2_schedule_swap to ensure the buffers won't be destroyed
before the vblank event is dealt with.

This parallels the approach taken by the Intel DDX in commit
0d2392d44aae95d6b571d98f7ec323cf672a687f.

Fixes: http://bugs.freedesktop.org/show_bug.cgi?id=29065

v2: Don't write completion events to the client if it has quit.
v3: Don't try to unref the NULL buffers from a DRI2_WAITMSC event.
    Take a ref in schedule_swap earlier, so the offscreen fallback
    doesn't incorrectly destroy the buffers.

Signed-off-by: Christopher James Halse Rogers <christopher.halse.rogers@canonical.com>
Signed-off-by: Dave Airlie <airlied@redhat.com>