radeon/kms: fix output enumeration
If a system has a mix of DVI-I, and DVI-D connectors, or a mix
of HDMI-A and HDMI-B connectors, the enumeration was per
connector type, but the naming was the same for multiple
connectors (DVI-[0-n] for DVI-I, -D, -A, and HDMI-[0-n]
for HDMI-A, and -B), you'd end up with multiple connectors
with the same name.  Fix this by tracking the number of
connectors across similar connector types.