ddx/ati: Add option "SwapbuffersWait" to control vsync of DRI2 swaps.
A new optional kms driver option "SwapbuffersWait" is defined
for xorg.conf, which defaults to "on". If "on", DRI2 bufferswaps
will be synchronized to vsync, otherwise not.

This currently only affects copy-swaps, not pageflipped swaps.
It also requires a swap_interval setting of zero by the OpenGL
client.

Ideally, we'd provide a way for dri2 to pass the current swap
interval to the ddx so we could change this dynamically.

Signed-off-by: Mario Kleiner <mario.kleiner@tuebingen.mpg.de>