Suggest where to find xorg-macros in case it's missing
[Modified version of U. Artie Eoff's commit to remove the duplicated
string. The paraphrased commit message is repeated below for clarity.]

xorg-macros is often package under an unusual name, so provide some
clues as to what name to try on the user's distribution. Installing
build-deps would be just too easy...

Also include a check for the existence of the XORG_DRIVER_CHECK_EXT
macro, otherwise configure might continue in spite of the missing macro
and cause obtuse syntax errors.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>