Properly restore CLOCK_CNTL_INDEX register
RADEONRestore() restores the CLOCK_CNTL_INDEX register before calling
RADEONRestorePLLRegisters(), which modifies the CLOCK_CNTL_INDEX register
to access the PLL registers.  As a result we may end up with the wrong clock
being selected when exiting X.  This happens on platforms where the driver
doesn't save and restore the VGA state.

Signed-off-by: Mark Kettenis <kettenis@openbsd.org>