sna: Tweak retiring old bo
As we check for retirement everytime we wakeup, it is seldom useful to
check again until we know we have invoked an operation that may block.
But when we do check, we do not want to scan the entire active list
looking for flushing candidates, so track those on a separate list.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>