sna: Move an assertion to the point of import
All the guesswork is so that when we require an inactive bo, we do
actually get a buffer that is not currently on a GPU active list. For
some unresolved reason, this assertion was firing when putting the
buffer onto the inactive list - so just workaround the worrisome issue
by delaying the check until use.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>