sna: Move the source to the GPU if it is reused
We attempt to skip upload a source pixmap to the GPU in the event it is
used only once (for example during image upload by firefox). However, if
we continue to use the CPU source pixmap then it obviously was worth
uploading to the GPU. So if we use the CPU pixmap a second time, do the
upload and then blit.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>