sna: Self-intersection of wide PolyLine are only drawn once
We need to process the union of the PolyLine command if lineWidth!=0 so
we cannot generally feed lineWidth==1 into our special case handler.
Proving the lines do not intersect is as difficult as finding the
intersections and thereby finding the union of the path - so there is no
advantage in adding a check whether a wide line could be special cased.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>