sna: Tweak the rendering priorities
If the last operation was on the GPU, continue on the GPU if this
operation overlaps any GPU damage or does not overlap CPU damage.
Otherwise, if the last operation was on the CPU only switch to the GPU
if we do not overlap any CPU damage and overlap existing GPU damage.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>